Solve Day 09: Sensor Boost

+++ src/main/kotlin/edu/mou/adventofcode/day05SunnyChanceAsteroids/Day05SunnyChanceAsteroids.kt
@@ -3,17 +3,17 @@ package edu.mou.adventofcode.day05SunnyChanceAsteroids
 import edu.mou.adventofcode.tools.DayProblem
 import edu.mou.adventofcode.tools.intcode.IntCodeInterpreter
 
-class Day05SunnyChanceAsteroids : DayProblem<Int, Int>(day = "05", name = "Sunny with a Chance of Asteroids") {
-    override fun solvePart1(vararg args: String): Int {
+class Day05SunnyChanceAsteroids : DayProblem<Long, Long>(day = "05", name = "Sunny with a Chance of Asteroids") {
+    override fun solvePart1(vararg args: String): Long {
         return runInterpreter(*args)
     }
 
-    override fun solvePart2(vararg args: String): Int {
+    override fun solvePart2(vararg args: String): Long {
         return runInterpreter(*args)
     }
 
-    private fun runInterpreter(vararg args: String): Int {
-        val interpreter = IntCodeInterpreter(input.first(), args.map { it.toInt() }.toMutableList())
+    private fun runInterpreter(vararg args: String): Long {
+        val interpreter = IntCodeInterpreter(input.first(), args.map { it.toLong() }.toMutableList())
         interpreter.run()
         return interpreter.output().last()
     }
