+++ src/main/kotlin/edu/mou/adventofcode/day07AmplificationCircuit/Amplifier.kt
@@ -2,10 +2,10 @@ package edu.mou.adventofcode.day07AmplificationCircuit
 
 import edu.mou.adventofcode.tools.intcode.IntCodeInterpreter
 
-data class Amplifier(val code: String, val phaseSetting: Int) {
+data class Amplifier(val code: String, val phaseSetting: Long) {
     private val interpreter = IntCodeInterpreter(code, mutableListOf(phaseSetting), pauseOnEmptyInput = true)
 
-    fun run(signal: Int = 0): Int {
+    fun run(signal: Long = 0): Long {
         interpreter.input(listOf(signal))
         interpreter.run()
         return interpreter.output().last()
