+++ src/main/kotlin/edu/mou/adventofcode/day07AmplificationCircuit/AmplifierCircuit.kt
@@ -1,8 +1,8 @@
 package edu.mou.adventofcode.day07AmplificationCircuit
 
 data class AmplifierCircuit(val amplifiers: List<Amplifier>) {
-    fun run(): Int {
-        var nextSignal = 0
+    fun run(): Long {
+        var nextSignal = 0L
 
         while (!amplifiers.last().isHalted()) {
             for (amplifier in amplifiers) {
