+++ src/main/kotlin/edu/mou/adventofcode/day07AmplificationCircuit/Day07AmplificationCircuit.kt
@@ -2,25 +2,25 @@ package edu.mou.adventofcode.day07AmplificationCircuit
 
 import edu.mou.adventofcode.tools.AocMath
 import edu.mou.adventofcode.tools.DayProblem
-import java.lang.Integer.max
+import kotlin.math.max
 
 class Day07AmplificationCircuit(inputFile: String = "input.txt") :
-    DayProblem<Int, Int>(day = "07", name = "Amplification Circuit", inputFile = inputFile) {
+    DayProblem<Long, Long>(day = "07", name = "Amplification Circuit", inputFile = inputFile) {
 
-    override fun solvePart1(vararg args: String): Int {
+    override fun solvePart1(vararg args: String): Long {
         return runAmplifiers(args[0].toInt()..args[1].toInt())
     }
 
-    override fun solvePart2(vararg args: String): Int {
+    override fun solvePart2(vararg args: String): Long {
         return runAmplifiers(args[0].toInt()..args[1].toInt())
     }
 
-    private fun runAmplifiers(phaseSettings: IntRange): Int {
+    private fun runAmplifiers(phaseSettings: IntRange): Long {
         val program = input.first()
-        var higherSignal = Int.MIN_VALUE
+        var higherSignal = Long.MIN_VALUE
 
         for (inputValues in AocMath.permutations(phaseSettings)) {
-            val signal = AmplifierCircuit(inputValues.map { Amplifier(program, it) }).run()
+            val signal = AmplifierCircuit(inputValues.map { Amplifier(program, it.toLong()) }).run()
 
             higherSignal = max(higherSignal, signal)
         }
