+++ src/test/kotlin/edu/mou/adventofcode/Day02The1202ProgramAlarmTest.kt
@@ -4,7 +4,7 @@ import edu.mou.adventofcode.day02The1202ProgramAlarm.Day02The1202ProgramAlarm
 import edu.mou.adventofcode.tools.DayProblemTest
 
 class Day02The1202ProgramAlarmTest :
-    DayProblemTest<Day02The1202ProgramAlarm, Int, Int>(
+    DayProblemTest<Day02The1202ProgramAlarm, Long, Long>(
         dayProblem = Day02The1202ProgramAlarm::class,
         expectedPart1 = 3500,
         expectedPart2 = 910,
