+++ src/test/kotlin/edu/mou/adventofcode/Day07AmplificationCircuitTest.kt
@@ -4,7 +4,7 @@ import edu.mou.adventofcode.day07AmplificationCircuit.Day07AmplificationCircuit
 import edu.mou.adventofcode.tools.DayProblemTest
 
 class Day07AmplificationCircuitTest :
-    DayProblemTest<Day07AmplificationCircuit, Int, Int>(
+    DayProblemTest<Day07AmplificationCircuit, Long, Long>(
         dayProblem = Day07AmplificationCircuit::class,
         expectedPart1 = 65210,
         argsPart1 = listOf("0", "4"),
